Dropbear ssh: client config file
Adding support for a per-client configuration file. Akin to the one
provided by the OpenSSH client. This initial implementation is
minimalistic with just basic support for a few options.
* Default config file location is ~/.ssh/dropbear_config. Can't be
customized. Not using `config` to avoid incompatibilities or worse -
silent changes of bahvior in the ssh clients.
* Added DROPBEAR_DEFAULT_USE_SSH_CONFIG complie directive. Currently
the default is to build the client without config file support. Override
in your localoptions.h
* Manpage update.
* .gitignore the .vscode folder - Visual Studio Code private files.

tjk :)