Dropbear SSH test
On my local machine the dbclient is timing out unable to connect to the spawned
dropbear service. It is waiting for me to provide password. Even if I do so
it doesn't like it. Must be missing something.

What I found out is invalid -p value when launching dropbear service. It passes
an IP address in place of a port.

tjk :)