Remove SO_LINGER
It could cause channels to take up to 5 seconds to close(), which would block
the entire process. On busy TCP forwarding sessions this would result in
channels seeming stuck and new connections not being accepted.

We don't need to monitor for flushing failures since we can't report errors, so
SO_LINGER wasn't useful.

Thanks to GektorUA for reporting and testing

Fixes #230