dbclient: fix -y and save hostkey to known_hosts
The -y should behave as -oStrictHostKeyChecking=accept-new.
So we don't just accept new keys but store them to recheck if they weren't changed.

Without this the -y is same as -y -y if the know_hosts is empty.

The bug fix may be breaking if some script by mistake expected that the -y is just to disable the checking at all.