sysoptions.h: move keepalive string above ident
OpenWrt package removes Dropbear version from ident string.
Until now, that was done with sed, instead of a patch.

I'm trying to change that to a patch and I noticed that the version string
is defined only 3 lines above the ident string. A normal patch would
include that line. Any version change would invalidate the patch.

I moved the keepalive string just above the ident string,
as these are somewhat related. That will keep the version string outside
the patch.

Signed-off-by: Marius Dinu <m95d+git@psihoexpert.ro>